requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-restful==0.3.10
flask-jwt-extended==4.5.3
pydantic==2.5.2
python-dotenv==1.0.0 